requirements.txt
Django<1.9
django-cms<3.3
django-treebeard>=2.0
django-filer>=0.9.12
django-newsletter>=0.6

cmsplugin-filer>=0.10.2
cmsplugin-iframe==0.1.2
djangocms-admin-style>=1.0.6
djangocms-column
djangocms-flash
djangocms-file==0.1
djangocms-googlemap
djangocms-inherit
djangocms-link
djangocms-style
djangocms-text-ckeditor>=2.8.1
djangocms-picture
djangocms-teaser
djangocms-video
pytz
django-reversion>=1.8.7,<1.9
django-classy-tags>=0.3.4.1
html5lib
Pillow>=2.3
django-sekizai>=0.7
django-bootstrap-form>=3.2
six

# assuming MemCache and Postgres on the production system by default, so adding deps for it, but the choice
# is ultimately yours
psycopg2
python-memcached
